validation: move TxPos/Ord from contract modul
