+++ src/contract/seal.rs
@@ -21,9 +21,7 @@
 // limitations under the License.
 
 use core::fmt::Debug;
-use std::cmp::Ordering;
 use std::hash::Hash;
-use std::num::NonZeroU32;
 
 use bp::dbc::Method;
 pub use bp::seals::txout::blind::{ChainBlindSeal, ParseError, SingleBlindSeal};
@@ -33,10 +31,10 @@ pub use bp::seals::SecretSeal;
 use bp::{dbc, Outpoint, Tx, Txid, Vout};
 use commit_verify::{mpc, Conceal};
 use single_use_seals::SealWitness;
-use strict_encoding::{StrictDecode, StrictDumb, StrictEncode, StrictType};
+use strict_encoding::{StrictDecode, StrictDumb, StrictEncode};
 
 use crate::contract::xchain::Impossible;
-use crate::{XChain, XOutpoint, LIB_NAME_RGB_COMMIT};
+use crate::{XChain, XOutpoint};
 
 pub type GenesisSeal = SingleBlindSeal<Method>;
 pub type GraphSeal = ChainBlindSeal<Method>;
@@ -176,69 +174,6 @@ impl<U: ExposedSeal> XChain<U> {
             })
             .ok_or(self)
     }
-}
-
-#[derive(Copy, Clone, PartialEq, Eq, Hash, Debug, Display)]
-#[derive(StrictType, StrictDumb, StrictEncode, StrictDecode)]
-#[strict_type(lib = LIB_NAME_RGB_COMMIT)]
-#[cfg_attr(
-    feature = "serde",
-    derive(Serialize, Deserialize),
-    serde(crate = "serde_crate", rename_all = "camelCase")
-)]
-#[display("{height}@{timestamp}")]
-// TODO: Move into validation::Logic or vm::Contract
-pub struct TxPos {
-    height: u32,
-    timestamp: i64,
-}
-
-impl TxPos {
-    pub fn new(height: u32, timestamp: i64) -> Option<Self> {
-        if height == 0 || timestamp < 1231006505 {
-            return None;
-        }
-        Some(TxPos { height, timestamp })
-    }
-
-    pub fn height(&self) -> NonZeroU32 { NonZeroU32::new(self.height).expect("invariant") }
-}
-
-impl PartialOrd for TxPos {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> { Some(self.cmp(other)) }
-}
-
-impl Ord for TxPos {
-    fn cmp(&self, other: &Self) -> Ordering { self.timestamp.cmp(&other.timestamp) }
-}
-
-/// RGB consensus information about the current mined height of a witness
-/// transaction defining the ordering of the contract state data.
-#[derive(Copy, Clone, PartialOrd, Ord, PartialEq, Eq, Hash, Debug, Display, From)]
-#[derive(StrictType, StrictDumb, StrictEncode, StrictDecode)]
-#[strict_type(lib = LIB_NAME_RGB_COMMIT, tags = order)]
-#[cfg_attr(
-    feature = "serde",
-    derive(Serialize, Deserialize),
-    serde(crate = "serde_crate", rename_all = "camelCase", untagged)
-)]
-// TODO: Move into validation::Logic or vm::Contract
-pub enum TxOrd {
-    #[from]
-    #[display(inner)]
-    OnChain(TxPos),
-
-    #[display("offchain@{priority}")]
-    OffChain { priority: u32 },
-
-    #[display("archived")]
-    #[strict_type(dumb)]
-    Archived,
-}
-
-impl TxOrd {
-    #[inline]
-    pub fn offchain(priority: u32) -> Self { TxOrd::OffChain { priority } }
 }
 
 pub type XWitnessTx<X = Impossible> = XChain<Tx, X>;
